Merged PR 3780650: Produce error when RSA-PKCS1 output buffer is too small
The RSA-PKCS1 decryption code uses a side-channel safe copy function to write the variable-sized output in constant time. The code didn't return an error when it was truncating the output. This fix adds that error, and rationalizes the error handling in the function.
Added test for all the error conditions of PKCS1.

Related work items: #23466960