Merged PR 3965002: Fix up CMake files after recent changes
This change updates the CMake files to fix compilation errors introduced by the recent test changes. It also fixes a number of warnings related to obsolete declaration styles and unsafe CRT functions, and adds the SymCryptDependencies repo as a submodule so that dependencies can be pulled in automatically rather than having to be copied manually.

Tested: compiled on Windows with CMake and Razzle (scbuild.cmd), ran generic unit test. I'm currently unable to pull the changes on Linux due to ADO issues, but I will verify Linux compilation once those issues are resolved.