Merged PR 3991172: Create Azure Pipelines YAML script for Windows build
This change adds a YAML script to enable building CMake for AMD64 Windows 10 in Azure Dev Ops. This build should occur automatically for all pull requests into master, and whenever master changes (whether due to a PR or a force push, etc).

Also removed the dependency on `ntintsafe.h`, which was not actually being used. (Maybe Changyu removed calls to its functions it when he did the Linux port?)