Merged PR 3995913: Fix Linux build issues and unit test segfault; update Azure pipeline
This change updates the CMakeFiles and fixes some code issues to get the Linux build working again. I also fixed a segmentation fault in the unit test that occurred due to a case-sensitive comparison being used to load resources, which led to one resource not being found. I'm not sure why that wasn't hit when testing on Linux previously.

Also update `azure-pipelines.yaml` to support the Linux build and test.