Merged PR 4043652: Enable CPU-specific optimizations for AMD64
This change enables CPU-specific optimizations for AMD64 on Windows. To compile with optimizations, specify the AMD64 toolchain file using `-DCMAKE_TOOLCHAIN_FILE=cmake-toolchain/windows-amd64.cmake`. If no toolchain file is specified, we will default to a generic, unoptimized build.

Also updated the Azure pipelines definition to use the new toolchain file, and to do an "out of source" build - i.e. the build is now performed in a subdirectory and all build artifacts will be placed there instead of being side-by-side with the source code.