Merged PR 4135452: Fix printtable buffer error by switching from deprectated strstream to stringstream
My bad, last time I didn't realize this was actually a strstream and not a stringstream. When using strstream, after the final s << item, you're supposed to put a null terminator as s.str() returns a c string. Or, as the old code was, when constructing the string use pcount to construct the string based on the exact number of characters input to the stream. By removing this, it then overran and we got the crazy values seen.
However, my fix is just to switch over to stringstream, as strstream has been deprecated since C++98. Stringstream doesn't have the terminator requirement and .str() returns a String and not a c string.