Merged PR 5578565: AES-GCM: Stitch AES-CTR and GHASH operations
Publishing a PR as I've got to a state where we have good improvements
across the board and want to have feedback on whether this is a good way
to structure the code changes.

+ Implement stitched functions for AES-GCM xmm/neon
+ For AES-GCM enc/dec part functions, there is still a reasonable amount
of shared code to handle partial blocks. This should be rarely hit,
and I've kept the logic in aes-default.c - we only dispatch to the
stitched functions when we have full blocks to both en/decrypt and
authenticate
+ For Xmm we are quite register constrained, so phrasing such that we
can use a maximum of 16 registers at a time is important (otherwise we
can easily lose any benefit from spilling/loading unnecessarily)
  + In the hot loop we have 8 registers containing AES-CTR state, 3
registers containing GHASH accumulators and have 5 registers
remaining for temporary values (round keys, multiplication
constants, byte reversal constant, data to authenticate)
  + In earlier attempts to phrase this I was trying to avoid having to
reload the data to authenticate (it always has to be in a variable
in AES-CTR as the ciphertext), but doing this means having to unroll
less and is not worth it - it is better to just (re)load data to
authenticate 1 block at a time in parallel with doing some number of
AES rounds
+ For Neon I noticed we have been doing a redundant EXT in the inner
loop of GHASH (we swap top and bottom halves of the same v register
twice to do byte reversal and then compute top half EOR bottom half
for karatsuba multiplication). Rephrased to avoid this.
  + Otherwise the phrasing is very similar to Xmm. Could investigate
avoiding the reloading of GHASH data as we have very low register
pressure given 32 ASIMD registers. Currently do not have access to a
machine which merges AES instruction pairs (only have access to a
TX2), so I'm not going to look into this too much right now.

+ Still to do:
  + Expose specific AesGcm encrypt/decrypt functions from SymCrypt which
further reduce overheads (can avoid some checks which are not avoidable when
only have an optimized partial gcm function)
  + Implement SymCryptAesGcmXXcryptStitchedYmm to use VAES / VPCLMULQDQ

Depending on feedback I can either include these todos in another iteration of this PR,
or complete this and create a separate PR to handle these cases.

Related work items: #31417357, #31417376