Merged PR 6868565: Strip symbols for release builds of OE and embedded modules
On some systems, symbols are stripped from binaries at package install time. This is incompatible with our FIPS integrity verfication approach; our modules must be stripped before being run through the FIPS post-processing script or the the runtime integrity verification will not work. Although the symbol table is not in a loadable segment and is therefore not part of our FIPS boundary, stripping symbols changes size and offset information stored in the ELF header, which is within the FIPS boundary and therefore affects the result of the integrity check.

To avoid this issue but still minimize binary size, this change adds steps to the CMake build files to strip the modules if the build is a Release build. We keep a small subset of symbols which are required by the FIPS post-processing script; these symbols must not be stripped at install time either.