Merged PR 6872685: Ignore relocations in data section. Do not optimize Debug mode binaries.
It is possible for relocations to occur within the .data section, but that section is excluded from our FIPS boundary and therefore also from the writeable segment copy that the integrity verification code uses to reverse the module's relocations. However, the code didn't account for this, which resulted in reading and writing off the end of the buffer when relocations existed in .data, causing heap corruption. The fix is to just ignore these relocations since they're outside the FIPS boundary anyway.

Also changed CMakeFiles to set -O0 on Debug builds so that temporary values aren't optimized out. This might make the tests take longer to run, but it significantly improves debuggability.