Merged PR 6931593: Split debug symbols using objcopy before stripping
This allows us to preserve the symbols separately from the stripped binary, so that they can be loaded separately in GDB, or uploaded to a symbol server. See [objcopy documentation](https://www.man7.org/linux/man-pages/man1/objcopy.1.html) for more info on how this works.

Also added `--build-id` linker flag which stores a unique identifier in the ELF file. The ID can be used to map preserved symbols to the corresponding stripped binary.