Merged PR 7775773: Fix integer truncation issue in 32-bit Linux builds
This change fixes an incorrect implementation of the `SYMCRYPT_MUL32x32TO64` when compiling for 32-bit architectures using gcc or clang. Previously the macro cast the input parameters to `unsigned long`, which is 32 bits on some platforms and 64 bits on others. This resulted in the multiplication truncating the top half of the word. The fix is to cast to UINT64 instead.

Also added 32-bit Linux unit tests to the Azure pipeline so it will be built and tested in our CI builds.

Related work items: #32615192