Merged PR 8129085: Disable AVX2 in Windows boot lib for AMD64
We're investigating boot failures in the lab (see bugs #42178021, #42182776) but so far we've made limited progress because the boot debugger unexpectedly disconnects, and we haven't been able to reproduce the issue locally. Our working theory is that recent SHA2 optimizations may have increased code size and/or stack usage beyond what the boot environment can accommodate. To confirm or exclude this diagnosis, this change disables AVX2 for AMD64 in the boot environment, which will remove the new SHA2 optimizations.

Related work items: #42178021, #42182776