Merged PR 8880994: Unpin Windows build container version
We previously pinned the Windows build container to a specific version (SHA-256 hash) to work around a regression in the container image that broke Python. That regression has long since been fixed, and the image version we were pinned to is no longer cached, which was causing Windows CI builds to take a very long time to initialize. This change puts us back on the latest container version.