Merged PR 9299172: Rename package.json to SymCryptPackage.json
`package.json` is used by NPM and causes automated tooling to flag our repo as missing required files for NPM projects. Of course, we don't use NPM and our `package.json` is used to configure which build outputs are included in which SymCrypt packages. The simple fix is to just rename the file.