Merged PR 9617187: Add /Gy flag when building with MASM
Due to an internal linker issue, MASM files must be assembled with the `/Gy` flag, or linking fails for some Windows components. `/Gy` causes each function to be in its own "section". The linker can only eliminate entire sections at a time, so when many functions are in the same section, if one of them is alive, all of them get emitted. This should match existing Razzle behavior.

This only applies to x86 and AMD64.

Related work items: #42154581