Merged PR 9631977: Use /Gy for ClCompile on all architectures
The previous changes to add the `/Gy` flag for MASM and ClCompile (C/C++) only applied to x86 and AMD64. For ARM64 ASM, `/Gy` is not supported by the assembler, so I assumed this was also true for the compiler. But it turns out that was a faulty assumption; `/Gy` is both supported and required for ARM64.

Note that the x86 and AMD64 assemblers (MASM) use a different target in the project file than the ARM64 assembler (MARMASM), so we can just apply `/Gy` globally to both MASM and ClCompile targets.

Related work items: #42154581