Merged PR 9784893: Enable v2 unwind data
This change adds the `-d2epilogunwindrequirev2` flag to ClCompile. This flag was previously set automatically by Razzle but is not currently set by the undocked LKG compiler. V1 unwind data means that the unwinder needs to read the code to unwind epilogues, which make cause a page fault if the unwind operation is happening at raised IRQL and the code page is paged out. Hence the unwinder stops when encountering V1 unwind data and a machine frame was just unwound (since machine frames can interrupt a routine in its epilogue).

Related work items: #47272287