Merged PR 9883477: Fix issues on arm32 linux
## Description:

This PR configures ARM32 memory accesses to access each byte individually for potentially unaligned accesses via the SYMCRYPT_LOAD_ and SYMCRYPT_STORE_ macros because some ARM32 CPUs require 64 bit accesses to be aligned on 32 bit boundaries (e.g. SONIC's ARM32 test environment).

When building for ARM32 we explicitly link with libc and libgcc because this fixes some load time symbol resolution issues. This also causes the emitted ELF segments to be ordered differently so we change the integrity verification to account for that.

## Admin Checklist:
- [ ] You have updated documentation in symcrypt.h to reflect any changes in behavior
- [ ] You have updated CHANGELOG.md to reflect any changes in behavior
- [ ] You have updated symcryptunittest to exercise any new functionality
- [ ] If you have introduced any symbols in symcrypt.h you have updated production and test dynamic export symbols (exports.ver / exports.def / symcrypt.src) and tested the updated dynamic modules with symcryptunittest
- [ ] If you have introduced functionality that varies based on CPU features, you have manually tested with and without relevant features
- [ ] If you have made significant changes to a particular algorithm, you have checked that performance numbers reported by symcryptunittest are in line with expectations
- [ ] If you have added new algorithms/modes, you have updated the status indicator text for the associated modules if necessary