Merged PR 10076162: Rename SymCrypt.sys to symcryptk.dll
The SymCrypt kernel mode module is not technically a driver in that it does not interface with hardware devices, does not create a device object, and in fact does not use any WDF functionality. As such, we've been advised by the kernel team that it should use the .dll extension instead of .sys, as e.g. ci.dll does. It turns out that this is not just a convention, but actually a requirement, because certain system components assume that .sys files will always be found under `%SystemRoot%\System32\drivers`, but we don't want to put SymCrypt there (since it's not a driver).

We still build the project with the `WindowsKernelModeDriver10.0` toolset to get access to KMDF headers and libraries.