Merged PR 10088810: Fix incorrect file copy in package JSON
An error in the JSON file for package.py caused the symcryptk.lib import library to overwrite the symcryptk.dll implementation library. This change fixes the error. Also re-enables CheckCFlags, as the reason it was failing was due to it correctly recognizing that the import library is not a valid PE file.