Merged PR 10090086: Fix CMake pipeline build
Due to a [known compiler issue](https://developercommunity.visualstudio.com/t/Illegal-Instruction-AVX512/10511452), the CMake build is producing an executable with AVX512 instructions in places they shouldn't be, which caused a crash due to illegal instruction execution on machines that don't support this instruction set. As a temporary workaround, we need to roll the CMake pipeline back to the previous container version.