Merged PR 10177217: Remove outdated security advice on RC4 comment
Per [this GitHub issue](https://github.com/microsoft/SymCrypt/issues/34), the advice on key generation for RC4 is outdated and no longer best practice. RC4 should never be used except for backwards compatibility, but this part of the comment is unhelpful in any case, so this PR removes it.

Thanks to [Myriachan](https://github.com/Myriachan) for reporting this issue.