Merged PR 10534989: Fix CallBackAlloc to adjust the nBytes in case the size is not aligned.
## Description:
SymCryptCallBackAlloc failed as part of a call to SymCryptRsakeyGenerate with the new SymCrypt.
After investigation, the failure occurs since we changed the API from "posix_alloc" to "aligned_alloc".

aligned_alloc - expect to get the number of bytes to allocate as an integral multiple of the alignment.

So, when passing not a multiple of the alignment number the function failed.

I suspect this issue is important since the regular Linux product use the API without the check of the size alignment.
Hence, I created this PR with a fix that adjust the size we receive to be aligned.

## Admin Checklist:
- [ ] You have updated documentation in symcrypt.h to reflect any changes in behavior
- [ ] You have updated CHANGELOG.md to reflect any changes in behavior
- [ ] You have updated symcryptunittest to exercise any new functionality
- [ ] If you have introduced any symbols in symcrypt.h you have updated production and test dynamic export symbols (exports.ver / exports.def / symcrypt.src) and tested the updated dynamic modules with symcryptunittest
- [ ] If you have introduced functionality that varies based on CPU features, you have manually tested with and without relevant features
- [ ] If you have made significant changes to a particular algorithm, you have checked that performance numbers reported by symcryptunittest are in line with expectations
- [ ] If you have added new algorithms/modes, you have updated the status indicator text for the associated modules if necessary

Fix CallBackAlloc to adjust the nBytes for allocation in case the size is not aligned

Signed-off-by: v-shlevy <v-shlevy@microsoft.com>

Related work items: #49419416