Merged PR 11124474: Always run daily build even when there is no churn
Currently the daily build only runs if there is churn in the branch, which is a bit pointless since it's automatically kicked off by each completed PR anyway. The daily build is also intended to prevent us from being broken by container/tools changes (which has happened before), so we need to run it even if there is no churn. This PR fixes that.