Merged PR 11163422: Wipe temporary values used by integrity test
## Description:

FIPS 140-3 TE05.08.02 requires that "any temporary values generated during the integrity test are zeroised upon completion of the integrity test." We were not previously doing this, because the temporary values are not secrets, but it is required for certification. This PR adds the required zeroisation.

## Admin Checklist:
- [X] You have updated CHANGELOG.md to reflect any changes in behavior
(Others N/A)

Related work items: #52849999