Merged PR 11660319: Change MSBuild architecture to x64
The newest OneBranch container uses a new WDK version which now requires certain tools to use the same architecture as the host. This caused our Windows undocked build pipelines to break because MSBuild defaults to using the x86 binaries, even on AMD64 hosts. This resulted in an ApiValidation error:

```
##[error]ApiValidation(0,0): Error : While Checking file C:\__w\1\s\build\bin\amd64chk\kmdll\symcryptk.dll aitstatic is returned exit code 193.
    13>ApiValidation : error : While Checking file C:\__w\1\s\build\bin\amd64chk\kmdll\symcryptk.dll aitstatic is returned exit code 193. [C:\__w\1\s\modules\windows\kernel\symcryptk.vcxproj]
```

Setting `msbuildArchitecture: 'x64'` in the `VSBuild` task fixes the issue.