Merged PR 13274908: Add CMake presets
This is the first of many build refactoring changes that I'm working on to facilitate the creation of NuGet packages and potentially other multi-architecture packaging solutions for SymCrypt. The goal is to move the information about supported configurations out of the Azure DevOps pipeline YAML and build.py script, to the greatest extent possible. This will make it easier to refactor the ADO pipeline YAML into different stages, and will also make it easier for external consumer to quickly see which configurations are supported (e.g. via `cmake --list-presets`).

Tested: used presets file to build all supported configurations on Linux and Windows. On macOS, the ARM64 presets work, but I'm currently not able to successfully cross-compile from ARM64 to x86-64 (whether using presets or not).