Merged PR 13534590: Rust AES-GCM: Implement GHash
This is a first draft of my Rust implementation of GHash as part of AES-GCM. I wanted to send it out early so that I can start getting feedback on interface design, use of idiomatic Rust, which parts of the C code we want to hew closely to versus which parts we want to substantially change, and generally code style.

To allow for smaller reviews and more rapid feedback, I'm planning to check in the code in small-ish pieces like this, checking each one into `feature/rust_aes_gcm` until the end-to-end is complete. Then I will merge that branch into `feature/verifiedcrypto`. I think this will be preferable to having one large review at the end, but let me know if I end up going too far in the other direction and spamming everyone with many small reviews.