Merged PR 13623094: Improve build documentation/ease of use
This PR adds documentation for how to build SymCRust, and improves ease of use for the developer, ensuring that targets (test, bench and normal build) can be run on the codebase without changing build files. It updates the `CMakeLists.txt` files for the C static libraries to use the default "lib" prefix on Unix-likes per convention, as Cargo expects this and cannot be forced to not look for it; this allows the build to succeed on Unix-like platforms.

Tested on Windows and Ubuntu: `cargo build`, `cargo test --features std`, `cargo bench --features benchmarking`, CMake build with `-DSYMCRYPT_SYMCRUST=ON`