Merged PR 13588380: Add accessor functions for `extern const` data symbols
Exporting data symbols from a dynamic module makes linking complicated on Windows. It requires the headers to define the symbols as `__declspec(dllimport)` only on the consuming side, which requires either two separate headers, or special macros. Without `declspec(dllimport)` the linker will not be able to find the data symbols. This has tripped up a number of consumers.

Providing accessor functions alleviates this issue, and is better practice for a dynamic library anyway, because it provides better ABI stability and encapsulation.

Background:
- [Import into an application using __declspec(dllimport)](https://learn.microsoft.com/en-us/cpp/build/importing-into-an-application-using-declspec-dllimport?view=msvc-170)
 - [LNK4217 - Russ Keldorph's WebLog - Site Home - MSDN Blogs (archive.org)](https://web.archive.org/web/20140911111411/http://blogs.msdn.com/b/russellk/archive/2005/03/20/399465.aspx?wa=wsignin1.0)

Related work items: #52850075