Merged PR 13727530: Include SymCrypt undocked properties to fix SDL errors
This change fixes binskim error BA2008 which was caused by not having control flow guard enabled for the NuGet test application. Although that application is only meant to test the NuGet package and thus is never shipped anywhere, it's easier to just fix it than try to get an exception.

Related work items: #58731016