Merged PR 14172824: Enable ASM for macOS
macOS build and test was taking an inordinately long time in our CI pipelines due to the lack of ASM support, and overall resulting in a poor development experience for macOS users. This change adds ASM support to macOS on Apple Silicon with the following changes:

- Removes `.type` directive from function prologue, which is not supported on Mach-O
- Updates symbol mangling to support macOS, which requires `_` prefix
- Adds support for CPU feature detection using macOS APIs
- Updates CMake Presets and pipeline YAML to enable ASM
- Removes redundant static unit tests from macOS build pipeline (dynamic test already tests both static + dynamic)

Intel Macs still do not support ASM as there is additional work required to build for that target successfully.

Note that we still don't consider macOS officially supported for production scenarios, but this PR removes the last major blocker for being able to add that support in the future, if business requirements dictate it.

Tested: CI pipeline, local build and test on Apple Silicon Mac