Merged PR 14302331: Fix perf measurement on Linux ARM64 systems
As of Ubuntu 24.04, `clock_gettime` is no longer a reliable way to measure performance at the granularity we need for our testing infrastructure. It did work at some point, although always knew that it probably wasn't as accurate as we'd like. But in testing AES-GCM perf today, I found that the perf tests would run for an inordinately long time because `calibratePerfMeasurements` was calculating a minimum measurement clock time of ~50k cycles, which caused the scaling logic to set `g_fixedTimeLoopsRuns` to 512, resulting in every execution of the fixed time loop taking a very long time.

Changing `GET_PERF_CLOCK()` to directly read `cntvct_el0` fixes this. The minimum measurement time is now 1000, matching Windows, and the sanity check values are in their expected ranges.

Tested on ARM64 Surface device